fix : unit test in generate router
