+++ pkg/generator/testdata/internal/controllers/storage/v1/sc_list/storage.go
@@ -1,8 +1,6 @@
-package storages
+package sclist
 
-import (
-	"github.com/sev-2/raiden"
-)
+import "github.com/sev-2/raiden"
 
 type MyStorage struct {
 	raiden.BucketBase
@@ -16,4 +14,10 @@ func (r *MyStorage) Name() string {
 }
 func (r *MyStorage) Public() bool {
 	return true
+}
+
+type StorageController struct {
+	raiden.ControllerBase
+	Http    string `path:"/sc-list" type:"storage"`
+	Storage MyStorage
 }
